+++ packages/better-define/tests/fixtures.test.ts
@@ -1,12 +1,12 @@
 import { resolve } from 'node:path'
 import {
   rollupBuild,
-  RollupEsbuildPlugin,
   RollupJson,
   RollupNodeResolve,
   RollupVue,
   RollupVueJsx,
   testFixtures,
+  UnpluginOxc,
 } from '@vue-macros/test-utils'
 import { describe } from 'vitest'
 import VueBetterDefine from '../src/rollup'
@@ -21,9 +21,7 @@ describe('fixtures', async () => {
         RollupVueJsx(),
         RollupJson(),
         RollupNodeResolve(),
-        RollupEsbuildPlugin({
-          target: 'esnext',
-        }),
+        UnpluginOxc.rollup(),
       ]),
     {
       cwd: resolve(__dirname, '..'),