+++ packages/jsx-directive/src/core/v-for.ts
@@ -67,7 +67,9 @@ export function resolveVFor(
               s,
               0,
               'renderList',
-              version ? 'vue' : '@vue-macros/jsx-directive/helpers',
+              lib.startsWith('vue')
+                ? 'vue'
+                : '@vue-macros/jsx-directive/helpers',
             )
         return `${renderList}(${list}, ${params} => `
       }
@@ -84,14 +86,17 @@ export function transformVFor(
 ): void {
   if (nodes.length === 0) return
 
-  nodes.forEach(({ node, attribute, parent, vMemoAttribute }) => {
+  nodes.forEach(({ node, attribute, parent, vIfAttribute, vMemoAttribute }) => {
     const hasScope = ['JSXElement', 'JSXFragment'].includes(
       String(parent?.type),
     )
-    s.prependRight(node.end!, `)${hasScope ? '}' : ''}`)
+    s.prependRight(
+      node.end!,
+      `)${hasScope ? (vIfAttribute ? '' : '}') : '}</>'}`,
+    )
     s.appendLeft(
       node.start!,
-      `${hasScope ? '{' : ''}${resolveVFor(attribute, node, s, options, vMemoAttribute)}`,
+      `${hasScope ? (vIfAttribute ? '' : '{') : '<>{'}${resolveVFor(attribute, node, s, options, vMemoAttribute)}`,
     )
     s.remove(attribute.start! - 1, attribute.end!)