+++ packages/jsx-directive/src/core/v-html.ts
@@ -1,10 +1,11 @@
+import type { OptionsResolved } from '..'
 import { isVue2, type JsxDirective } from '.'
 import type { MagicStringAST } from '@vue-macros/common'
 
 export function transformVHtml(
   nodes: JsxDirective[],
   s: MagicStringAST,
-  version: number,
+  { version }: OptionsResolved,
 ): void {
   nodes.forEach(({ attribute }) => {
     s.overwriteNode(