+++ packages/macros/tests/fixtures.test.ts
@@ -1,10 +1,10 @@
 import { resolve } from 'node:path'
 import {
   rollupBuild,
-  RollupEsbuildPlugin,
   RollupVue,
   RollupVueJsx,
   testFixtures,
+  UnpluginOxc,
 } from '@vue-macros/test-utils'
 import { describe } from 'vitest'
 import VueMacros from '../src/rollup'
@@ -25,9 +25,7 @@ describe('fixtures', async () => {
             vueJsx: RollupVueJsx(),
           },
         }),
-        RollupEsbuildPlugin({
-          target: 'esnext',
-        }),
+        UnpluginOxc.rollup(),
       ])
     },
     {